+++ linmod/utils.py
@@ -13,8 +13,26 @@ def expand_grid(**columns):
     return df.sort(columns.keys())
 
 
+def pl_crps(samples_column: str, truth_column: str):
+    """
+    Monte Carlo approximation to the CRPS.
+    """
+
+    samples = pl.col(samples_column)
+    truth = pl.col(truth_column)
+    n = samples.len()
+
+    return (samples - truth).abs().mean() - 0.5 * (
+        samples.head(n - 1) - samples.tail(n - 1)
+    ).abs().mean()
+
+
 def pl_softmax(pl_expr, over=None):
     if over is None:
         return pl_expr.exp() / pl_expr.exp().sum()
     else:
         return pl_expr.exp() / pl_expr.exp().sum().over(over)
+
+
+def pl_mae(samples_column: str, truth_column: str):
+    return (pl.col(samples_column) - pl.col(truth_column)).abs().mean()
